TEZ-4721: Drop Hadoop 2.x support profiles and specific shims implementation
- Retains the core HadoopShim framework and GuavaShim
- Removes hadoop27 and hadoop28 profiles from all pom.xml files
- Removes hadoop-shim-impls module (Hadoop 2.7 and 2.8 specific shims)
