+++ tez-plugins/pom.xml
@@ -29,37 +29,15 @@
   <artifactId>tez-plugins</artifactId>
   <packaging>pom</packaging>
 
-  <profiles>
-    <profile>
-      <id>hadoop27</id>
-      <activation>
-        <activeByDefault>false</activeByDefault>
-      </activation>
-      <modules>
-        <module>tez-protobuf-history-plugin</module>
-        <module>tez-yarn-timeline-history</module>
-        <module>tez-yarn-timeline-history-with-acls</module>
-        <module>tez-history-parser</module>
-        <module>tez-aux-services</module>
-      </modules>
-    </profile>
-    <profile>
-      <id>hadoop28</id>
-      <activation>
-        <activeByDefault>true</activeByDefault>
-      </activation>
-      <modules>
-        <module>tez-protobuf-history-plugin</module>
-        <module>tez-yarn-timeline-history</module>
-        <module>tez-yarn-timeline-history-with-acls</module>
-        <module>tez-yarn-timeline-cache-plugin</module>
-        <module>tez-yarn-timeline-history-with-fs</module>
-        <module>tez-history-parser</module>
-        <module>tez-aux-services</module>
-      </modules>
-    </profile>
-
-  </profiles>
+  <modules>
+    <module>tez-protobuf-history-plugin</module>
+    <module>tez-yarn-timeline-history</module>
+    <module>tez-yarn-timeline-history-with-acls</module>
+    <module>tez-yarn-timeline-cache-plugin</module>
+    <module>tez-yarn-timeline-history-with-fs</module>
+    <module>tez-history-parser</module>
+    <module>tez-aux-services</module>
+  </modules>
 
   <build>
     <plugins>
